Move relations into "SQL" namespace
