+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/event_repository.rb
@@ -1,4 +1,4 @@
-require 'ruby_event_store/index_violation_detector'
+require 'ruby_event_store/rom/sql/index_violation_detector'
 require 'ruby_event_store/rom/unit_of_work'
 
 module RubyEventStore
