+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/sql.rb
@@ -1,0 +1,2 @@
+require_relative 'sql/relations/events'
+require_relative 'sql/relations/stream_entries'
