+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/sql/index_violation_detector.rb
