+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/sql/rake_task.rb
@@ -1,3 +1,4 @@
 require 'rom/sql/rake_task'
 require 'ruby_event_store/rom'
+require 'ruby_event_store/rom/sql'
 load 'ruby_event_store/rom/tasks/migration_tasks.rake'
