+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/sql/relations/stream_entries.rb
