+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/sql/tasks/migration_tasks.rake
@@ -1,3 +1,4 @@
+require 'ruby_event_store/rom'
 
 desc 'Setup ROM EventRespository environment'
 task 'db:setup' do
